WIP: Gather secrets for targeted OSP services
This patch introduces the ability to gather secrets for the available
namespaces.
We're limiting secrets gather only to the defined OpenStack services,
because some components like nova and cinder encode the whole config
file in a secrets and don't rely anymore on the initContainer.

Note this is still a WIP as we might want to improve filtering and
apply masking on the base64 decoded data.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>