Get neutron status from the existing ctlplane
This patch improves the way we're gathering information from the
existing ctlplane services. In particular, for each service it's
possible to define an entry in the get_status() switch/case and
point to the implemented service function. As an example, this
patch introduces the same commands we're running for neutron
through SOS reports [1]. The output is saved within the ctlplane
directory generated by the must-gather tool.

[1] https://github.com/sosreport/sos/blob/main/sos/report/plugins/openstack_neutron.py#L57-L64

Signed-off-by: Francesco Pantano <fpantano@redhat.com>