Gather data from kuttl namespaces
This patch introduces a new, generic function to be able to expand the
existing NAMESPACES related array with an arbitrary number of elements
that reflect a certain criteria. In particular, the gather script adds
a new line that grows the existing array adding every project that
matches the kuttl keyword: by doing this we can finally run the regular
resource gathering against this set of additional namespaces.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>