Return the output of the run
In some cases when we run the must-gather customers will see errors in
the output such as:

  error: the server doesn't have a resource type "nad"

And when they attach the must-gather directories and send them to us we
will not have this information, which can prove useful.

To resolve this issue this patch changes how we run the must-gather so
that we can save the contents of the output in a run.log that gets
included in the output directory of the must-gather.

To do this we move the contents of the `gather` script to one called
`gather_run` which we call from the new `gather` scripts.