Add debug mode
When we are developing scripts the whole flow is a bit cumbersome
because we have to write the code changes, build the container, upload
the container, run the new container, and repeat until we get things
right.

This patch documents a possible development workflow that leverages a
new script called `gather_debug` that allows us to go into the container
and work there, then copy the changes we've done in the container, and
let the `oc adm must-gather` command complete as normal.