Trigger Guru Meditation Reports
Most OpenStack services have the possibility of having Guru
Meditation Reports: https://github.com/openstack/oslo.reports

These reports are presented in the logs, so we have to trigger them
before we gather the logs.

This patch adds the mechanims to triger those reports in whichever way
each service has configured: Via signal or via file.

The example provided is for the cinder service, which is configured to
use files because the services are running in the host PID namespace, so
we don't know the PID of the service.