Per container logging
We are currently gathering the logs for all the containers in a pod into
the same file making it very hard to read since the individual logs are
concatenated and not interlaced by time.

With this patch we now have a log file per pod container, making it
easier to read them.