Gather SOS reports
This patch adds SOS gathering for controller nodes so we get all
information with the must-gather request and the reports are gathered
together instead of needing to got to 4 different places.

SOS gathering is done in parallel for all hosts, and they they are
stored uncompressed in the must-gather report to avoid having nested
compression that makes it more difficult to use it.

Since we may not always be interested in gathering SOS reports or we may
want just SOS reports of nodes for specific services we have the
`SOS_SERVICES` environmental variable that we can use to define the
services.

To speed things up we don't gather a full SOS report but instead limit
the plugins used to `block,cifs,crio,devicemapper,devices,iscsi,lvm2,
memory,multipath,nfs,nis,nvme,podman,process,processor,selinux,scsi,udev`.
A user can change this using the `SOS_ONLY_PLUGINS` environmental
variable.