Extract shellcheck configuration from Makefile
This patch extracts the shellcheck configuration from the Makefile where
we are adding things in the command line invocation and moves them to
the .shellcheckrc file.

This patch also adds the location of the sources to the configuration so
we no longer need to add the "shellcheck disable=SC1091" directive
before sourcing files to prevent errors.

Some editors that do not use the .shellcheckrc may still report the
SC1091 error on those lines, but the `make check` will not, since it
uses the .shellcheckrc file.