New parallel processing mechanism
This patch adds a new parallel processing mechanism with 2 functions:
- run_gb: To run commands in background
- wait_bg: Wait for backgroup commands to complete

The number of simultaneous concurrent processes are controlled by the
environmental variable CONCURRENCY, which defaults to 5.

The patch moves the SOS report gathering that is already running things
in the backgroud to use this new mechanism.

To ensure that the concurrency is honored we need to source the
gather_sos file instead of calling it from the command line, so the
gather_sos has been modified to support sourcing and calling mechanisms.

The calling mechanism is useful when debugging things.