Parallelize gather_ctlplane_resources
Move the gather_ctlplane_resources to the new parallel processing
mechanism by sourcing the file from gather_run instead of calling it and
adapting its code to support both sourcing and calling mechanisms.

The calling mechanism is useful when debugging things.

This patch also removes the nested loop that made multiple `oc`
requests, now it makes a single request to gather the
pod-node-previouslog tuple of information.

It also fixes a bug that I introduced in an earlier patch regarding the
gathering of previous container logs.