Add SOS report of EDPM nodes
This patch adds a mechanism to gather SOS reports from EDPM nodes.

Unlike the SOS of controller nodes, this mechanism is disabled by
default so that openstack-must-gather doesn't generate SOS reports for
any EDPM node.

Using environmental variable `SOS_EDPM` we can request gathering SOS
reports for all or specific EDPM nodes:

- SOS_EDPM=all ==> Gathers SOS reports for ALL nodes. Useful for CI and
  testing.
- SOS_EDPM=edpm-compute-0,edpm-compute-1 ==> Gathers SOS reports for
  those 2 sepecific nodes. Useful for example for issues with live
  migration.

Just like the SOS gathering of controller nodes, the SOS reports are
stored uncompressed in the must-gathere report to facilitate analysis.

This patch assumes there's connectivity between the debug container and
the EDPM nodes, which at least is the case for CRC deployments with
install_yamls. If we find that that's not good enough we'll have to
change the approach to use Pods/Jobs with the right networks.

Time it took to run the openstack-must-gather gathering SOS reports for
the CRC and 2 EDPM nodes was around 8m19s, which is around 40 seconds
more that without gathering the EDPM nodes.

Once the PR that adds the `openstack_edpm` profile to `sos` [1] merges
we have to add it to the default list of profiles.

[1]: https://github.com/openstack-k8s-operators/openstack-must-gather/pull/18