Use a different variable to collect CRS
This patch fixes in the first place the usage of the "resources" variable.
The purpose of this variable is to define a set of k8s related services
that should be collected for each namespace; for this reason it's
defined in common.sh and exported, so it can be used in any script that
is sourced by "gather_run".
Instead of using the same "$resources" variable that is used a lot in the
collection-scripts, this change customizes the variable name used to collect
CR(s) and apiservices, and it avoids collisions with other 'get_resources'
happening in other namespaces.
In addition, this patch fixes the way CRD(s) are collected: they are
global and shouldn't be collected "per namespace", hence they're now
collected within the "crd/" folder created in the root must-gather directory.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>