Optionally dump the openstack databases
This patch introduces the ability to specify an environment variable
with a list of openstack databases that should be dumped. It is also
possible to set the variable to "ALL", which results in trigger the
mysqldump command with the --all-databases option.
By default the OPENSTACK_DATABASES variables is unset and this flow
is entirely skipped. To enable it, the aforementioned variable should
be explicitly set.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>