Add dump-conf option to the mask secret script
This patch adds the ability to the mask script to detect a .conf file (as
key in a secret), and dump its content in the current directory. This
helps to ease the troubleshooting when looking for a service config file:
instead of base64 -d a portion of the secret, we provide a dump of that
file in the form <secret.yaml>-<config_file_name>.conf within the same
directory.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>