Use controlplane ip from compute nodes instead of hostname for gathering the sosreport logs
This patch is a followup of #49 since we can't ssh to $hostname from
must-gather container because it doesn't have access to OSP controlplane
DNS server.

To enhance readability and maintainability of the script, acquiring the
required parameters in the $data variable was changed from
go-template to json, using jq command as its parser.

An updated Dockerfile was provided to install jq in the must-gather
image.

Signed-off-by: Roberto Alfieri <ralfieri@redhat.com>