Update gather_sos* to not decompress collected sos reports
Having decompressed sos reports is breaking the Zuul web interface for users, due to how Zuul works the front end downloads a manifest file and uses that to generate the artifacts and logs tab for the build.

Our jobs are creating 20mb+ manifest files and these need to be downloaded and processed. The file size isn't the main issue, the file count in the manifest file causes high processing and is locking up users browsers.

Just stopping the compression on sos reports halves the file count in the manifest file.

This patch adds support for passing `SOS_DECOMPRESS` environment
variable to disable decompressing sos reports. Default is to to allow
decompressing.