SOS Report: Get rotated container logs
Current must gather only gathers the logs in the pod's current log file
and the previous container run, so once the logs are rotated we no
longer have access to that information through the `oc` command and that
information is not present in the must gather tar.xz.

Rotated logs are only available under `/var/log/pods` and, since log
rotation is not an OpenShift responsibility, there is no way to get them
besides accessing that directory.

There is no sos plugin that gathers the contents of `/var/log/pods` only
for `/var/log/containers`, but in this last directory there are only
symlinks to the last files, not the rotated ones.

I see 3 ways to solve this issue:

- Submit a PR to upstream SOS report repository with a new plugin to
  gather that directory. This would take time and we would need to wait
  until it gets merged and released upstream, then released downstream,
  and a new container is built using the latest release.

- Create a custom toolbox container image based on the RHEL9 one that
  has our custom SOS report plugin and use it in openstack-must-gather.
  This means that we would have to productize another container.

- Just compress and transfer manually the whole directory.

This patch takes the last of these approaches, as it is fast and easy to
implement and is not affected by the size limits that logs in sos
reports usually have.

The approach creates a tar.xz in the same debug container as the sos
report and then both are transferred simultaneously using the same debug
container.

Since the new tar.xz file will contain all the logs we no longer need to
run all those individual `oc logs` commands we were running before.

Jira: https://issues.redhat.com/browse/OSPRH-7615