SOS report: Only create and modify dir when needed
We recently added the option to not decompress the sos report tar.xz
file, but in that case we are still creating a directory that is going
to be empty and we make a call to modify its access.

This patch changes it so that the directory is only created when it's
going to be used and the same for changing the permissions.