SOS report: Separated download debug logs
For convenience when downloading the sos report we are requesting extra
debug information from the debug session, but this information is only
necessary if the operation fails.

That's why in this patch we separate the debug logs (stderr) and the
actual data download (stdout) and delete the debug logs when the
download is successful.