Cleaner logs on EDPM SOS reports
When we are downloading the SOS reports from EDPM nodes, the whole tar
data is shown in stdout, making them a mess.

In this patch we change how we get the SOS from the nodes, and instead
of using `tee`, we redirect stdout to the file and stderr to stdout, so
we still see any errors on the download but we no longer have to see the
whole tar in stdout.