SOS: Gather /var/log including journal
In OSP17 and earlier our SOS reports for the control plane included the
`/var/log` directory including the journal files that can be read (using
all filtering options) using `journalctl -D`.

This information is not present in our current SOS reports from the
control plane, which is problematic for example to debug a lot of the
storage attaching issues.

This patch changes our default plugins to include `/var/log` and passes
the `--all-logs` in the `sos report` call to ensure the journal is also
included.