EDPM SOS: Gather journal
In OSP17 and earlier our SOS reports for the compute nodes included the
journal in the `/var/log` directory.

Our current sos reports for EDPM nodes don't include this information,
because it's only present if the `--all-logs` flag is passed to the `sos
report` command.

This patch ensures we are passing the `--all-logs` flag when gathering
sos reports from EDPM nodes.