[SOS EDPM] Use openstack_edpm profile
The `openstack_edpm` plugin was being gathered on the OCP cluster nodes,
but it's intended for the EDPM nodes, so we remove it from the list of
plugins in `SOS_ONLY_PLUGINS`.

We move the `openstack_edpm` to the EDPM sos gathering script replacing
the old `openstack` profile that is not as appropriate for
openstack-operator deployments.