SOS: Gather nodes where operators are running
This patch adds a new configuration boolean flag using env var named
`SOS_OSP_OPERATORS` to gather sos reports for nodes that are running the
operators.

This is important when customers must send SOS reports of a deployment
that couldn't get to the point of deploying services.

By default openstack-must-gather will get the SOS reports where the OSP
operators are running.