OCP SOS: Only gather rotated logs
To remove the duplicated logs, using `oc logs` and via sos reports, this
patch changes how sos reports gather logs and limits them to only
rotated logs.

We are forced to ignore tar errors because a file can be removed while
processing it due to the log rotation mechanism.

The code only adds the logs if there are any rotated logs available.