Symlink to rotated logs
We currently have 2 locations that have container logs:

- `must-gather/namespaces/*/pods/*/logs/`
- `must-gather/sos-reports/_all_nodes/*/var/log/pods/*/...`

This means that we could inadvertently miss some of the files if we
don't know of the two locations.

To alleviate this situation this patch builds symbolic links to the
rotated log files in the `namespaces` logs location.

Because there can be 2 log files with the same name on 2 different
nodes, the code renames them.