Fix: Code is not waiting for background tasks
Our current `gather_run` script doesn't wait until all background tasks
have completed when the database is not being gathered.

This is caused by the `gather_db` script using `exit 0` instead of
`return` when sourced.

This patch makes that small change and now the code will wait for all
background operations as intended.

This can be problematic in some cases because we can find the
must-gather tar file missing some of the data that we want collected.