Produce a single archive
This patch adds a step at the end of the gathering to produce an
additional single archive of the whole gathered content.

Use case: I need to grep across pod logs with a request ID so I need to
download the content of the gathering. However the content is
in a lot of small files causing that a recursive wget takes tens of
minutes to finish and probably generates excess load on the log server.