SOS OCP: Add crypto plugin
Add the `crypto` plugin to the list of plugins gathered by the SOS
report on the OCP nodes.

This is necessary for example to gather FIPS enabled status on the
nodes.

There's no need to do anything for EDPM nodes because we gather the
`system` profile there and the `crypto` plugin is part of that profile.