Do not try to get logs from non existent pods
This patch fixes the current issue related to the fact that we try to
get logs from pods that are not present in the namespace. This is caused
by the fact that the crash_status variable is sometimes not predictable
and an output made by many strings is produced. We only need to know if
the Pod was terminated, and get the -previous log.
In addition, given the complexity generated by go-template, we moved the
data processing to json and jq, as it results easy to read and maintain.

Jira: https://issues.redhat.com/browse/OSPRH-8197

Signed-off-by: Francesco Pantano <fpantano@redhat.com>
(cherry picked from commit dbbc8b9c73d5c5dc1483cf753806343cde020611)