Skip collecting logs from pods if they are pending
If a pod is in Pending state and has no 'status.containerStatus' field,
the jq command crashes, and we get no logs. This change sets the
container name for those pods as null, and then we can skip collecting
logs for that pod, while still calling 'oc describe' on it.