Gather glance task information
This patch set's up the glance directory for service info gathering.
It also adds the task list which tells us about the status of
import operations.

Future work:
In future, it would be useful to add stores info which tells us
about configured backends and their properties but it is not
currently available in openstackclient.