Add OS_GIT_VERSION to the openstack-must-gather build system
As per the guidelines [1], a must-gather image must produce a version
file that indicates the product and the version, following the
"major.minor.micro.qualifier" approach. By doing this is possible to
integrate it with other tools that need this information to process the
associated data.
Patch [2] produces a "version" file as part of the must-gather
execution, while here both the Makefile and the Containerfile are
modified to support this new ENV var.

[1] https://github.com/openshift/must-gather/blob/master/must-gather.md#must-gather-images
[2] https://github.com/openstack-k8s-operators/openstack-must-gather/pull/81

Signed-off-by: Francesco Pantano <fpantano@redhat.com>