Add per namespace database dump and cell db filtering
We currently filter cellX databases due to the hardcoded "| grep cell".
This prevents collecting information from the nova cells dedicated
galera instances.
This patch improves this aspect and provides a FILTER_DB_CELLS env var
set to "false" by default. By doing this we're able to keep backward
compatibility with the previous behavior, but also do not always exclude
cells DBs.
However, OPENSTACK_DATABASES env var is global, so the dump_db function
is improved to check whether a service is present or not, and gracefully
skip the database dump if it is not found in the current galera instance.
Because the main loop iterates over the namespaces (we used to run a
"oc get openstackcontrolplane --all-namespaces ..."), the databases
dumps are now collected in a per-namaspaces path like the following:
<must-gather>/dbs/<namespace>/<galera_pod>-<service>.sql. This is valid
for both --all-databases and selective dumps.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>