Fix ConfigMap masking not applied when called from gather_run
The masking code in gather_services_cm was only executed when the script
was run directly (CALLED=1), but not when sourced and called from
gather_run. This caused passwords and sensitive data in ConfigMaps to
remain unmasked during must-gather collection.

Move the wait_bg and masking logic inside the gather_services_cm()
function itself, following the same pattern used in gather_secrets.
This ensures masking is applied regardless of how the script is invoked.

Follow-up to: 466f4e78 (Mask sensitive data in CRs and ConfigMaps)

Jira: OSPRH-20621

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>